# docs/swagger.yaml
basePath: /v1
definitions:
  entity.Lint:
    properties:
      organization_id:
        example: 1
        type: integer
      project_id:
        example: 1
        type: integer
      rule:
        example: json string
        type: string
    type: object
  v1.response:
    properties:
      error:
        example: message
        type: string
    type: object
  v1.uploadRequest:
    properties:
      organization_id:
        example: 1
        type: integer
      project_id:
        example: 1
        type: integer
      rule_yml:
        example: json string
        type: string
    required:
    - organization_id
    - project_id
    - rule_yml
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: Go Clean Template API
  version: "1.0"
paths:
  /lint/rule/{project_id}:
    get:
      consumes:
      - application/json
      description: get project's lint rule
      operationId: getRule
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Lint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: get project's lint rule
      tags:
      - lint
  /lint/upload:
    post:
      consumes:
      - application/json
      description: create or update project's lint rule
      operationId: upload
      parameters:
      - description: Upload Rule
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.uploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: create or update project's lint rule
      tags:
      - lint
swagger: "2.0"